requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.1
django-extensions==3.2.3
django-stubs==5.0.4
django-stubs-ext==5.0.4
mysql-connector-python==9.0.0
mysqlclient==2.2.4
sqlparse==0.5.1
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.1
